requirements.txt
Flask
PyPDF2
SQLAlchemy
pdfplumber
camelot-py
python-docx
opencv-python
pyyaml
requests